requirements.txt
anyascii==0.3.1
asgiref==3.5.2
beautifulsoup4==4.9.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
decorator==5.1.1
Django==4.0.6
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
fonttools==4.37.1
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jieba==0.42.1
kiwisolver==1.4.4
l18n==2021.3
matplotlib==3.5.3
numpy==1.23.2
openpyxl==3.0.10
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
ratelim==0.1.6
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
urllib3==1.26.11
wagtail<= 5.0.1
webencodings==0.5.1
Willow==1.4.1
wordcloud==1.8.2.2
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
